requirements.txt
streamlit
ultralytics
open_clip_torch
faiss-cpu
pillow
opencv-python
tqdm
torch
